bye bye proto.Call
also moved Batch-specific functionality to proto/batch.go
(from api.go); BatchRequest is still a Request, but that
is bound to change.
